refactor: decompose 4 more monolith files into submodules (issue #2486 sprint 4)
- interactive_manipulation.py (919 → 162 LOC) → _mouse_picking, _ik_solver,
  _constraint_manager, _pose_library
- learners.py (933 → 16 LOC) → _base, _bc (behavioral cloning), _dagger, _gail
- steam_engine.py (919 → 410 LOC) → _constants, _models, _property_backends,
  _vapor_pressure
- acid_gas_dewpoint_calculator.py (926 → 60 LOC) → _acid_gas_calculator,
  _acid_gas_models, _acid_gas_utils, _acid_gas_widget

All public symbols remain importable from original module paths.
