+++ src/shared/python/upstream_drift_tools/calculators/thermo/_constants.py
@@ -1,0 +1,57 @@
+from __future__ import annotations
+
+STANDARD_ATMOSPHERIC_PRESSURE: float = 101325.0
+BOILING_TEMPERATURE_WATER: float = 373.15
+LIQUID_WATER_PRESSURE_THRESHOLD: float = 101325.0
+SPECIFIC_GAS_CONSTANT_WATER: float = 461.5
+LIQUID_WATER_SPECIFIC_HEAT: float = 4186.0
+VAPOR_ENTHALPY_REFERENCE: float = 2500.0
+VAPOR_ENTHALPY_SLOPE: float = 1.9
+VAPOR_SPECIFIC_HEAT_CP: float = 1.9
+VAPOR_SPECIFIC_HEAT_CV: float = 1.4
+
+VAPOR_ENTROPY_REFERENCE: float = 8000.0
+VAPOR_ENTROPY_SLOPE: float = 2000.0
+
+FAHRENHEIT_TO_CELSIUS_OFFSET: float = 32
+FAHRENHEIT_TO_CELSIUS_SCALE: float = 5 / 9
+CELSIUS_TO_FAHRENHEIT_SCALE: float = 9 / 5
+
+MIN_TEMPERATURE_UI_K: float = 273.15
+MAX_TEMPERATURE_UI_K: float = 647.15
+
+NEWTON_RAPHSON_TOLERANCE: float = 1e-6
+NEWTON_RAPHSON_DERIVATIVE_TOLERANCE: float = 1e-10
+NEWTON_RAPHSON_MAX_ITERATIONS: int = 10
+NEWTON_RAPHSON_STEP_SIZE: float = 0.1
+
+ANTOINE_A: float = 8.07131
+ANTOINE_B: float = 1730.63
+ANTOINE_C_CELSIUS: float = 233.426
+ANTOINE_C_KELVIN: float = 39.724
+
+BUCK_A: float = 6.1121
+BUCK_B: float = 18.678
+BUCK_C: float = 234.5
+BUCK_D: float = 257.14
+
+MMHG_TO_PASCAL_FACTOR: float = 133.322
+PASCAL_TO_MMHG_FACTOR: float = 0.00750062
+MBAR_TO_KPA_FACTOR: float = 10.0
+KPA_TO_PA_FACTOR: float = 1000.0
+KELVIN_TO_CELSIUS_OFFSET: float = 273.15
+
+CRITICAL_TEMPERATURE_WATER: float = 647.15
+CRITICAL_PRESSURE_WATER: float = 22.064e6
+TRIPLE_POINT_TEMPERATURE: float = 273.16
+TRIPLE_POINT_PRESSURE: float = 611.657
+
+DEFAULT_DEW_POINT_TEMPERATURE_CELSIUS: float = 25.0
+DEFAULT_QUALITY: float = 0.5
+
+FALLBACK_ATMOSPHERIC_PRESSURE: float = 101325.0
+FALLBACK_BOILING_TEMPERATURE: float = 373.15
+
+SUPERHEATED_STATE: int = 0
+SATURATED_FROM_TEMP_STATE: int = 1
+SATURATED_FROM_PRESSURE_STATE: int = 2
